LayoutTests/fast/history/page-cache-notification-showing-expected.txt

Tests that a page with a notification that is showing is able to enter the back/forward cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the back/forward cache
PASS Received closed notification
PASS restoredFromCache is true
PASS successfullyParsed is true

TEST COMPLETE
